Add tooltip initialization and region dropdown population functionality
