+++ html/scripts/main.js
@@ -4,6 +4,8 @@ import { setupPostalCodeClicks, loadSelectedPostalCodes, setMode, togglePostalCo
 import { setupModeToggle, setupLookupButton } from './uiSetup.js';
 import { connectWebSocket } from './websocket.js';
 import { setupLassoSelect } from './lasso.js';
+import { initializeTooltips } from './tooltip.js';
+import { populateRegionDropdown, handleRegionChange } from './regionManager.js';
 
 async function loadRegionOptions() {
     try {
@@ -28,104 +30,6 @@ async function loadRegionOptions() {
     }
 }
 
-async function populateRegionDropdown() {
-    try {
-        const response = await fetch('data/regions.json');
-        const regions = await response.json();
-        const select = document.getElementById('regions');
-
-        // Clear existing options
-        select.innerHTML = '';
-
-        regions.forEach(item => {
-            if (item.optgroup) {
-                // Create an optgroup
-                const optgroup = document.createElement('optgroup');
-                optgroup.label = item.optgroup;
-                
-                item.options.forEach(option => {
-                    const optionElement = document.createElement('option');
-                    optionElement.value = option.value;
-                    optionElement.textContent = option.text;
-                    // Store countries data as a data attribute
-                    if (option.countries) {
-                        optionElement.dataset.countries = JSON.stringify(option.countries);
-                    }
-                    optgroup.appendChild(optionElement);
-                });
-                
-                select.appendChild(optgroup);
-            } else {
-                // Create a regular option
-                const option = document.createElement('option');
-                option.value = item.value;
-                option.textContent = item.text;
-                select.appendChild(option);
-            }
-        });
-
-        // Set the initial selected option based on the current URL
-        const urlParams = new URLSearchParams(window.location.search);
-        const currentMap = urlParams.get('map');
-        // if (currentMap) {
-        //     const value = `/?map=${currentMap}`;
-        //     select.value = value;
-        //     handleRegionChange(select.selectedOptions[0]);
-        // }
-
-        // Add event listener to handle selection change
-        select.addEventListener('change', (event) => {
-            const selectedOption = event.target.selectedOptions[0];
-            handleRegionChange(selectedOption);
-        });
-    } catch (error) {
-        console.error('Error populating region dropdown:', error);
-    }
-}
-
-function handleRegionChange(selectedOption) {
-    if (!selectedOption) return;
-
-    // Get the current map from URL
-    const urlParams = new URLSearchParams(window.location.search);
-    const currentMap = urlParams.get('map');
-    
-    // Get the selected map value (strip the "/?map=" prefix)
-    const selectedMap = selectedOption.value.replace('/?map=', '');
-
-    // Only redirect if the map is actually different
-    if (selectedMap && selectedMap !== currentMap && !selectedOption.dataset.countries) {
-        window.location.href = selectedOption.value;
-        return;
-    }
-
-    // Otherwise, just toggle the countries without page reload
-    const countries = selectedOption.dataset.countries ? 
-        JSON.parse(selectedOption.dataset.countries) : [];
-
-    // Get all country toggles
-    const countryToggles = document.querySelectorAll('.country-item input[type="checkbox"]');
-    
-    // Toggle countries based on selection
-    countryToggles.forEach(toggle => {
-        const countryItem = toggle.closest('.country-item');
-        const countryName = countryItem.querySelector('span').textContent;
-        
-        // Check if the country should be visible for this region
-        const shouldBeVisible = countries.length === 0 || 
-            countries.some(c => countryName.toLowerCase().includes(c.toLowerCase()));
-        
-        // Only toggle if the state is different
-        if (toggle.checked !== shouldBeVisible) {
-            toggle.checked = shouldBeVisible;
-            toggleCountryVisibility(countryName, shouldBeVisible);
-        }
-    });
-
-    // Trigger zoom to visible after toggling the region
-    triggerZoomVisible();
-}
-
 function initializeApp() {
     loadSVG()
         .then(({ svgElement, originalViewBox }) => {
@@ -141,14 +45,14 @@ function initializeApp() {
             return loadSelectedPostalCodes();
         })
         .then(() => {
-            console.log('Postal codes loaded');
+            // console.log('Postal codes loaded');
             // ensureControlsVisibility();
         })
         .catch(error => console.error('Error initializing app:', error));
 }
 
 function initializeTooltip(element, tooltipText, placement) {
-    console.log('Initializing tooltip:', element, tooltipText, placement);
+    // console.log('Initializing tooltip:', element, tooltipText, placement);
     const tooltip = document.createElement('div');
     tooltip.className = 'tooltip';
     tooltip.textContent = tooltipText;
@@ -167,22 +71,22 @@ function initializeTooltip(element, tooltipText, placement) {
     });
 
     element.addEventListener('mouseenter', () => {
-        console.log('Mouse entered:', element);
+        // console.log('Mouse entered:', element);
         tooltip.style.visibility = 'visible';
         tooltip.style.opacity = '1';
         popperInstance.update().then(() => {
-            console.log('Popper updated:', tooltip);
+            // console.log('Popper updated:', tooltip);
         });
     });
 
     element.addEventListener('mouseleave', () => {
-        console.log('Mouse left:', element);
+        // console.log('Mouse left:', element);
         tooltip.style.visibility = 'hidden';
         tooltip.style.opacity = '0';
     });
 
     // Log to confirm event listeners are attached
-    console.log('Event listeners attached for:', element);
+    // console.log('Event listeners attached for:', element);
 }
 
 document.addEventListener('DOMContentLoaded', function() {
@@ -206,22 +110,7 @@ document.addEventListener('DOMContentLoaded', function() {
     });
 
     // Initialize tooltips
-    const tooltipElements = [
-        { element: document.getElementById('country-count'), tooltipText: 'Click to add individual countries', placement: 'top' },
-        { element: document.getElementById('region-label'), tooltipText: 'Select Regions of countries', placement: 'bottom' },
-        { element: document.getElementById('lookup-button'), tooltipText: 'Lookup in Mitoz', placement: 'top' },
-        { element: document.getElementById('zoom-in'), tooltipText: 'Zoom in', placement: 'top' },
-        { element: document.getElementById('zoom-out'), tooltipText: 'Zoom out', placement: 'top' },
-        { element: document.getElementById('zoom-factor'), tooltipText: 'Currend zoom', placement: 'top' },
-        { element: document.getElementById('reset-zoom'), tooltipText: 'Reset zoom to 1', placement: 'top' },
-        { element: document.getElementById('zoom-visible'), tooltipText: 'Zoom to visible area', placement: 'top' },
-        { element: document.getElementById('lasso-group'), tooltipText: 'Toggle lasso tool', placement: 'top' }
-    ];
-
-    tooltipElements.forEach(({ element, tooltipText, placement }) => {
-        console.log('Initializing tooltip for element:', element);
-        initializeTooltip(element, tooltipText, placement);
-    });
+    initializeTooltips();
 
     // Add event listener to close the lookup results
     document.querySelector('#lookup-results .lookup-close-btn').addEventListener('click', function() {
